requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.1
pydantic-settings==2.6.1

# Document parsing
pdfplumber==0.11.4
python-docx==1.1.2

# AI
openai==1.55.3

# Utilities
python-dotenv==1.0.1
